image.pollinations.ai/upscaler/requirements.txt
flask
torch
torchvision
numpy
pillow
requests
basicsr
realesrgan
gunicorn
werkzeug
